This is the homework 7
The app that I developed in this assignment is a “math quiz” app that teaches someone how to add two-digit numbers. It will display a pair of numbers and then the user needs to enter their sum and the app tells them whether they are correct. The user can select a difficulty level that determines how complicated the addition should be.